+++ packages/core/_src/global.ts
@@ -1,9 +1,9 @@
 declare global {
   /**
-   * @tsplus type fp-ts/data/ReadonlyArray
+   * @tsplus type fp-ts/core/ReadonlyArray
    * @tsplus type ReadonlyArray
    * @tsplus type Iterable
-   * @tsplus companion fp-ts/data/ReadonlyArray.Ops
+   * @tsplus companion fp-ts/core/ReadonlyArray.Ops
    * @tsplus companion ReadonlyArray.Ops
    */
   interface ReadonlyArray<T> {}