+++ packages/schema/_src/ext.ts
@@ -3,7 +3,7 @@ import { constant, pipe } from "@effect-app/core/Function"
 import * as NonEmptySet from "@effect-app/core/NonEmptySet"
 import type { ComputeFlat } from "@effect-app/core/utils"
 import { typedKeysOf } from "@effect-app/core/utils"
-import type { None, Some } from "@fp-ts/data/Option"
+import type { None, Some } from "@fp-ts/core/Option"
 import { v4 } from "uuid"
 
 import type { FromProperty } from "./_api.js"
@@ -511,7 +511,7 @@ export function makeRequired<NER extends Record<string, MO.AnyProperty>>(
 
 export function createUnorder<T>(): Order<T> {
   return {
-    compare: (_b: T) => (_a: T) => 0
+    compare: (_a: T, _b: T) => 0
   }
 }
 export function makeSet<ParsedShape, ConstructorInput, Encoded, Api>(