+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @effect-app/core
 
+## 0.12.1
+
+### Patch Changes
+
+- Disable trace
+
+## 0.12.0
+
+### Minor Changes
+
+- Update to effect and fp-ts 0.1 line
+
 ## 0.11.3
 
 ### Patch Changes