+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @effect-app/schema
 
+## 0.12.1
+
+### Patch Changes
+
+- Disable trace
+- Updated dependencies
+  - @effect-app/core@0.12.1
+
+## 0.12.0
+
+### Minor Changes
+
+- Update to effect and fp-ts 0.1 line
+
+### Patch Changes
+
+- Updated dependencies
+  - @effect-app/core@0.12.0
+
 ## 0.11.3
 
 ### Patch Changes