+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @effect-app/vue
 
+## 0.11.1
+
+### Patch Changes
+
+- Disable trace
+- Updated dependencies
+  - @effect-app/prelude@0.12.1
+  - @effect-app/schema@0.12.1
+  - @effect-app/core@0.12.1
+  - @effect-app/vue@0.11.1
+
+## 0.11.0
+
+### Minor Changes
+
+- Update to effect and fp-ts 0.1 line
+
+### Patch Changes
+
+- Updated dependencies
+  - @effect-app/prelude@0.12.0
+  - @effect-app/schema@0.12.0
+  - @effect-app/core@0.12.0
+  - @effect-app/vue@0.11.0
+
 ## 0.10.6
 
 ### Patch Changes